+++ .gitignore
@@ -323,3 +323,4 @@ $RECYCLE.BIN/
 
 .token
 .usertoken
+localca.crt
