+++ README.md
@@ -1,14 +1,15 @@
 # Home Assistant Twitch Server
 
-Home Assistant Twitch Server written on node. 
+Home Assistant Twitch Server written on node.
 
-This server forwards Twitch events to the home assistant server to trigger custom actions. 
+This server forwards Twitch events to the home assistant server to trigger custom actions.
 
 ## Prerequisites
 
 - node: https://nodejs.org/
 - yarn: https://yarnpkg.com/
 - running home assistant server: https://www.home-assistant.io/
+- node-red recommended: https://nodered.org/
 - Twitch dev account: https://dev.twitch.tv/
 
 ## Installation
@@ -18,3 +19,9 @@ TODO
 ## Usage
 
 TODO
+
+## Credits
+
+Initial code was based on: https://github.com/twitchdev/eventsub-webhooks-node-sample
+
+Also thanks a lot to all the developers of node, yarn, home assistant and node-red!
