+++ lib/data/scopes.js
@@ -1,3 +1,5 @@
+// https://dev.twitch.tv/docs/authentication#scopes
+
 module.exports = [
   'analytics:read:extensions', // View analytics data for the Twitch Extensions owned by the authenticated account.
   'analytics:read:games', // View analytics data for the games owned by the authenticated account.
