+++ lib/data/types.js
@@ -1,3 +1,5 @@
+// https://dev.twitch.tv/docs/eventsub/eventsub-subscription-types
+
 module.exports = [
   // 'channel.update', // Channel Update: A broadcaster updates their channel properties e.g., category, title, mature flag, broadcast, or language.
   'channel.follow', // Channel Follow: A specified channel receives a follow.
