+++ lib/main.js
@@ -3,59 +3,11 @@
 const fs = require('fs');
 const express = require('express');
 const bodyParser = require('body-parser');
-const http = require('http');
-const https = require('https');
 const crypto = require('crypto');
 const localtunnel = require('localtunnel');
 
-const paramsToUrl = (params) => Object.entries(params).map(param => `${param[0]}=${param[1]}`).join('&');
-
-const doRequest = (url, options, body) => new Promise(async (resolve) => {
-  let responseData = '';
-
-  const callback = (result) => {
-    result.setEncoding('utf8');
-    result.on('data', (d) => {
-      responseData += d;
-    }).on('end', () => {
-      if (!responseData) {
-        resolve();
-
-        return;
-      }
-
-      try {
-        const responseBody = JSON.parse(responseData);
-
-        resolve(responseBody);
-      } catch (e) {
-        resolve(responseData);
-      }
-    });
-  };
-
-  const bodyData = (body !== 'undefined') ? JSON.stringify(body) : '';
-
-  if (bodyData) {
-    if (!options.headers) options.headers = {};
-
-    options.headers['Content-Type'] = 'application/json';
-    options.headers['Content-Length'] = bodyData.length;
-  }
-
-  const request = (url.startsWith('http:'))
-    ? http.request(url, options, callback)
-    : https.request(url, options, callback);
-
-  request.on('error', (e) => {
-    console.log('');
-    console.error(e);
-  });
-
-  if (bodyData) request.write(bodyData);
-
-  request.end();
-});
+const request = require('./tools/request');
+const url = require('./tools/url');
 
 (async () => {
   const appSecret = process.env.SECRET;
@@ -129,7 +81,7 @@ const doRequest = (url, options, body) => new Promise(async (resolve) => {
       console.log('');
       console.log('Clear webhooks');
 
-      const resultGetWebhooks = await doRequest(`${twitchApiUrl}${helixEventsubSubscriptionPath}`, {
+      const resultGetWebhooks = await request(`${twitchApiUrl}${helixEventsubSubscriptionPath}`, {
         method: 'GET',
         headers: twitchApiAuthHeaders,
       });
@@ -149,7 +101,7 @@ const doRequest = (url, options, body) => new Promise(async (resolve) => {
 
         console.log(`- delete webhook ${webhookId}`);
 
-        await doRequest(`${twitchApiUrl}${helixEventsubSubscriptionPath}?id=${webhookId}`, {
+        await request(`${twitchApiUrl}${helixEventsubSubscriptionPath}?id=${webhookId}`, {
           method: 'DELETE',
           headers: twitchApiAuthHeaders,
         });
@@ -158,7 +110,7 @@ const doRequest = (url, options, body) => new Promise(async (resolve) => {
       }
     };
 
-    const createWebhook = (broadcasterId, type) => doRequest(`${twitchApiUrl}${helixEventsubSubscriptionPath}`, {
+    const createWebhook = (broadcasterId, type) => request(`${twitchApiUrl}${helixEventsubSubscriptionPath}`, {
       method: 'POST',
       headers: twitchApiAuthHeaders,
     }, {
@@ -181,7 +133,7 @@ const doRequest = (url, options, body) => new Promise(async (resolve) => {
     };
 
     // Get broadcaster id
-    const resultGetBroadcasterId = await doRequest(`${twitchApiUrl}${helixUsersPath}?login=${clientLogin}`, {
+    const resultGetBroadcasterId = await request(`${twitchApiUrl}${helixUsersPath}?login=${clientLogin}`, {
       method: 'GET',
       headers: twitchApiAuthHeaders,
     });
@@ -263,8 +215,8 @@ const doRequest = (url, options, body) => new Promise(async (resolve) => {
         return;
       }
 
-      await doRequest(haWebhookUrl, {
-        method: 'POST'
+      await request(haWebhookUrl, {
+        method: 'POST',
       }, event);
 
       console.log('');
@@ -314,7 +266,7 @@ const doRequest = (url, options, body) => new Promise(async (resolve) => {
       state: oauthState,
     };
 
-    res.redirect(`${twitchAuthUrl}/authorize?${paramsToUrl(params)}`);
+    res.redirect(url.make(`${twitchAuthUrl}/authorize`, params));
   });
 
   app.listen(port, async () => {
@@ -362,7 +314,7 @@ const doRequest = (url, options, body) => new Promise(async (resolve) => {
       scope: scopes.join('+'),
     };
 
-    const resultGetToken = await doRequest(`${twitchAuthUrl}/token?${paramsToUrl(params)}`, {
+    const resultGetToken = await request(url.make(`${twitchAuthUrl}/token`, params), {
       method: 'POST',
     });
